--- .editorconfig
@@ -1,141 +1,0 @@
-﻿[*.cs]
-
-# IDE0090: Utiliser 'new(...)'
-csharp_style_implicit_object_creation_when_type_is_apparent = false
-
-[*.cs]
-#### Styles de nommage ####
-
-# Règles de nommage
-
-dotnet_naming_rule.interface_should_be_begins_with_i.severity = suggestion
-dotnet_naming_rule.interface_should_be_begins_with_i.symbols = interface
-dotnet_naming_rule.interface_should_be_begins_with_i.style = begins_with_i
-
-dotnet_naming_rule.types_should_be_pascal_case.severity = suggestion
-dotnet_naming_rule.types_should_be_pascal_case.symbols = types
-dotnet_naming_rule.types_should_be_pascal_case.style = pascal_case
-
-dotnet_naming_rule.non_field_members_should_be_pascal_case.severity = suggestion
-dotnet_naming_rule.non_field_members_should_be_pascal_case.symbols = non_field_members
-dotnet_naming_rule.non_field_members_should_be_pascal_case.style = pascal_case
-
-# Spécifications de symboles
-
-dotnet_naming_symbols.interface.applicable_kinds = interface
-dotnet_naming_symbols.interface.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.interface.required_modifiers = 
-
-dotnet_naming_symbols.types.applicable_kinds = class, struct, interface, enum
-dotnet_naming_symbols.types.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.types.required_modifiers = 
-
-dotnet_naming_symbols.non_field_members.applicable_kinds = property, event, method
-dotnet_naming_symbols.non_field_members.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.non_field_members.required_modifiers = 
-
-# Styles de nommage
-
-dotnet_naming_style.begins_with_i.required_prefix = I
-dotnet_naming_style.begins_with_i.required_suffix = 
-dotnet_naming_style.begins_with_i.word_separator = 
-dotnet_naming_style.begins_with_i.capitalization = pascal_case
-
-dotnet_naming_style.pascal_case.required_prefix = 
-dotnet_naming_style.pascal_case.required_suffix = 
-dotnet_naming_style.pascal_case.word_separator = 
-dotnet_naming_style.pascal_case.capitalization = pascal_case
-
-dotnet_naming_style.pascal_case.required_prefix = 
-dotnet_naming_style.pascal_case.required_suffix = 
-dotnet_naming_style.pascal_case.word_separator = 
-dotnet_naming_style.pascal_case.capitalization = pascal_case
-csharp_style_expression_bodied_methods = false:silent
-csharp_style_expression_bodied_constructors = false:silent
-csharp_style_expression_bodied_operators = false:silent
-csharp_style_expression_bodied_properties = true:silent
-csharp_style_expression_bodied_indexers = true:silent
-csharp_style_expression_bodied_accessors = true:silent
-csharp_style_expression_bodied_lambdas = true:silent
-csharp_style_expression_bodied_local_functions = false:silent
-csharp_style_prefer_switch_expression = true:suggestion
-csharp_style_prefer_pattern_matching = true:silent
-csharp_style_pattern_matching_over_is_with_cast_check = true:suggestion
-csharp_style_pattern_matching_over_as_with_null_check = true:suggestion
-csharp_style_prefer_not_pattern = true:suggestion
-csharp_style_prefer_extended_property_pattern = true:suggestion
-csharp_prefer_simple_using_statement = true:suggestion
-csharp_prefer_braces = true:silent
-csharp_style_namespace_declarations = block_scoped:silent
-csharp_style_prefer_method_group_conversion = true:silent
-csharp_style_prefer_top_level_statements = true:silent
-csharp_style_prefer_primary_constructors = true:suggestion
-csharp_space_around_binary_operators = before_and_after
-csharp_indent_labels = one_less_than_current
-csharp_style_unused_value_assignment_preference = discard_variable:silent
-csharp_style_allow_blank_lines_between_consecutive_braces_experimental = true:silent
-csharp_style_prefer_local_over_anonymous_function = true:suggestion
-csharp_style_allow_blank_line_after_colon_in_constructor_initializer_experimental = true:silent
-csharp_style_allow_blank_line_after_token_in_conditional_expression_experimental = true:silent
-csharp_style_allow_blank_line_after_token_in_arrow_expression_clause_experimental = true:silent
-
-[*.vb]
-#### Styles de nommage ####
-
-# Règles de nommage
-
-dotnet_naming_rule.interface_should_be_commence_par_i.severity = suggestion
-dotnet_naming_rule.interface_should_be_commence_par_i.symbols = interface
-dotnet_naming_rule.interface_should_be_commence_par_i.style = commence_par_i
-
-dotnet_naming_rule.types_should_be_casse_pascal.severity = suggestion
-dotnet_naming_rule.types_should_be_casse_pascal.symbols = types
-dotnet_naming_rule.types_should_be_casse_pascal.style = casse_pascal
-
-dotnet_naming_rule.membres_autres_que_des_champs_should_be_casse_pascal.severity = suggestion
-dotnet_naming_rule.membres_autres_que_des_champs_should_be_casse_pascal.symbols = membres_autres_que_des_champs
-dotnet_naming_rule.membres_autres_que_des_champs_should_be_casse_pascal.style = casse_pascal
-
-# Spécifications de symboles
-
-dotnet_naming_symbols.interface.applicable_kinds = interface
-dotnet_naming_symbols.interface.applicable_accessibilities = public, friend, private, protected, protected_friend, private_protected
-dotnet_naming_symbols.interface.required_modifiers = 
-
-dotnet_naming_symbols.types.applicable_kinds = class, struct, interface, enum
-dotnet_naming_symbols.types.applicable_accessibilities = public, friend, private, protected, protected_friend, private_protected
-dotnet_naming_symbols.types.required_modifiers = 
-
-dotnet_naming_symbols.membres_autres_que_des_champs.applicable_kinds = property, event, method
-dotnet_naming_symbols.membres_autres_que_des_champs.applicable_accessibilities = public, friend, private, protected, protected_friend, private_protected
-dotnet_naming_symbols.membres_autres_que_des_champs.required_modifiers = 
-
-# Styles de nommage
-
-dotnet_naming_style.commence_par_i.required_prefix = I
-dotnet_naming_style.commence_par_i.required_suffix = 
-dotnet_naming_style.commence_par_i.word_separator = 
-dotnet_naming_style.commence_par_i.capitalization = pascal_case
-
-dotnet_naming_style.casse_pascal.required_prefix = 
-dotnet_naming_style.casse_pascal.required_suffix = 
-dotnet_naming_style.casse_pascal.word_separator = 
-dotnet_naming_style.casse_pascal.capitalization = pascal_case
-
-dotnet_naming_style.casse_pascal.required_prefix = 
-dotnet_naming_style.casse_pascal.required_suffix = 
-dotnet_naming_style.casse_pascal.word_separator = 
-dotnet_naming_style.casse_pascal.capitalization = pascal_case
-
-[*.{cs,vb}]
-tab_width = 4
-indent_size = 4
-dotnet_style_readonly_field = false:silent
-dotnet_diagnostic.IDE0051.severity = none
-dotnet_diagnostic.IDE0250.severity = none
-dotnet_diagnostic.IDE0251.severity = none
-dotnet_style_require_accessibility_modifiers = for_non_interface_members:silent
-dotnet_style_operator_placement_when_wrapping = beginning_of_line
-dotnet_code_quality_unused_parameters = all:silent
-dotnet_style_prefer_inferred_anonymous_type_member_names = true:suggestion
-dotnet_style_allow_multiple_blank_lines_experimental = true:silent
