+++ Assets/Prefabs/Board.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f2300f7f9b6293247975e138cbd5b66f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
