+++ Assets/Scripts/Board/GridElement.cs
@@ -1,16 +1,13 @@
 using UnityEngine;
 
-namespace Board
+public class GridElement
 {
-    public class GridElement
-    {
-        public Coord coord;
-        public GameObject gridElement;
+    public Coord coord;
+    public GameObject gameObject;
        
-        public GridElement(Coord coordVar, GameObject gridElementVar)
-        {
-            coord = coordVar;
-            gridElement = gridElementVar;
-        }
+    public GridElement(Coord coordVar, GameObject gameObjectVar)
+    {
+        coord = coordVar;
+        gameObject = gameObjectVar;
     }
 }
