--- Assets/Scripts/Board/Movement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40f99272864321745b258cd34caa7ef6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
