+++ Assets/Scripts/Gameplay/Ability.cs
@@ -19,6 +19,4 @@ public class Ability
         get { return _roundsBeforeReuse; }
         set { _roundsBeforeReuse = value; }
     }
-
-    public Ability() {}
 }
